Add new IMethod interface, create new DelegateMethod.
The new IMethod abstracts methods to a point where a method is more or
less a strategy, allowing for multiple methods that implement IMethod.
DelegateMethod does just this, however the instantiation parameters are
a bit different, requiring an agent, and task to delegate. Also added a
demo in world, delegating the clean task to agent 1.